requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albumentations==1.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.8.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
black==21.4b2
boto3==1.36.7
botocore==1.36.7
bpy==4.0.0
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
defusedxml==0.7.1
diffdist==0.1
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.9
einops==0.3.0
exceptiongroup==1.2.2
faiss-cpu
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.1.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iopath==0.1.9
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
kornia==0.5.0
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.3
lvis==0.5.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.23.0
omegaconf==2.1.1
open-clip-torch==2.0.2
openai==1.60.2
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
OpenEXR==3.3.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
Pillow==9.5.0
pip==24.0
platformdirs==4.3.6
portalocker==3.1.1
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyDeprecate==0.3.1
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.1
pyre-extensions==0.0.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3transfer==0.11.2
safetensors==0.5.2
scikit-image
scikit-learn
scipy
segment_anything==1.0
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==69.0.3
shapely==2.0.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
supervision==0.25.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
test_tube==0.7.5
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==0.9.2
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==1.13.1
torchmetrics==0.6.0
torchvision==0.14.1
tqdm==4.67.1
transformers==4.48.0
triton==3.2.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.4
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
wheel==0.42.0
wldhx.yadisk-direct==0.0.6
wrapt==1.17.2
xformers==0.0.16
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zstandard==0.23.0
